+++ android/src/main/java/me/carda/awesome_notifications/notifications/NotificationBuilder.java
@@ -1,26 +1,27 @@
 package me.carda.awesome_notifications.notifications;
 
 import android.app.Notification;
+import android.app.NotificationManager;
 import android.app.PendingIntent;
 
-import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.graphics.Bitmap;
 import android.graphics.Color;
-import android.media.AudioManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
+import android.service.notification.NotificationListenerService;
+import android.service.notification.StatusBarNotification;
 
 import com.github.arturogutierrez.BadgesNotSupportedException;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
+import java.util.Objects;
 
 import androidx.annotation.NonNull;
 import androidx.core.app.NotificationCompat;
@@ -28,11 +29,10 @@ import androidx.core.app.RemoteInput;
 
 import me.carda.awesome_notifications.AwesomeNotificationsPlugin;
 import me.carda.awesome_notifications.Definitions;
-import me.carda.awesome_notifications.R;
 import me.carda.awesome_notifications.notifications.broadcastReceivers.DismissedNotificationReceiver;
 import me.carda.awesome_notifications.notifications.broadcastReceivers.KeepOnTopActionReceiver;
 import me.carda.awesome_notifications.notifications.enumeratos.ActionButtonType;
-import me.carda.awesome_notifications.notifications.enumeratos.NotificationImportance;
+import me.carda.awesome_notifications.notifications.enumeratos.GroupSort;
 import me.carda.awesome_notifications.notifications.enumeratos.NotificationLayout;
 import me.carda.awesome_notifications.notifications.enumeratos.NotificationPrivacy;
 import me.carda.awesome_notifications.notifications.exceptions.PushNotificationException;
@@ -186,6 +186,8 @@ public class NotificationBuilder {
 
         NotificationCompat.Builder builder = new NotificationCompat.Builder(context, pushNotification.content.channelKey);
 
+        setGrouping(context, pushNotification, channel, builder);
+
         setVisibility(context, pushNotification, channel, builder);
         setShowWhen(pushNotification, builder);
 
@@ -212,8 +214,6 @@ public class NotificationBuilder {
         setLayoutColor(context, pushNotification, channel, builder);
 
         setBadge(context, channel, builder);
-
-        applyGrouping(channel, builder);
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             builder.setChannelId(channel.getChannelKey());
@@ -532,14 +532,35 @@ public class NotificationBuilder {
         }
     }
 
-    private void applyGrouping(NotificationChannelModel channelModel, NotificationCompat.Builder builder) {
+    private void setGrouping(Context context, PushNotification pushNotification, NotificationChannelModel channelModel, NotificationCompat.Builder builder) {
 
         if (!StringUtils.isNullOrEmpty(channelModel.groupKey)) {
             builder.setGroup(channelModel.groupKey);
 
-            if (BooleanUtils.getValue(channelModel.setAsGroupSummary)) {
+            boolean grouped = true;
+            if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.M) {
+
+                NotificationManager manager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
+                StatusBarNotification[] currentActiveNotifications = manager.getActiveNotifications();
+
+                for (StatusBarNotification activeNotification : currentActiveNotifications) {
+                    if (activeNotification.getGroupKey().contains("g:"+channelModel.groupKey)) {
+                        grouped = false;
+                        break;
+                    }
+                }
+            }
+
+            if (grouped) {
                 builder.setGroupSummary(true);
             }
+
+            String idText = pushNotification.content.id.toString();
+            String sortKey = Long.toString(
+                (channelModel.groupSort == GroupSort.Asc ? System.currentTimeMillis() : Long.MAX_VALUE - System.currentTimeMillis())
+            );
+
+            builder.setSortKey(sortKey + idText);
 
             builder.setGroupAlertBehavior(channelModel.groupAlertBehavior.ordinal());
         }