+++ lib/src/definitions.dart
@@ -1,6 +1,7 @@
 import 'package:awesome_notifications/src/enumerators/action_button_type.dart';
 import 'package:awesome_notifications/src/enumerators/default_ringtone_type.dart';
 import 'package:awesome_notifications/src/enumerators/group_alert_behaviour.dart';
+import 'package:awesome_notifications/src/enumerators/group_sort.dart';
 import 'package:awesome_notifications/src/enumerators/notification_importance.dart';
 import 'package:awesome_notifications/src/enumerators/notification_privacy.dart';
 import 'package:flutter/material.dart';
@@ -135,7 +136,7 @@ const NOTIFICATION_SOUND_PATH = 'sound';
 const NOTIFICATION_ENABLE_VIBRATION = 'enableVibration';
 const NOTIFICATION_VIBRATION_PATTERN = 'vibrationPattern';
 const NOTIFICATION_GROUP_KEY = 'groupKey';
-const NOTIFICATION_SET_AS_GROUP_SUMMARY = 'setAsGroupSummary';
+const NOTIFICATION_GROUP_SORT = 'groupSort';
 const NOTIFICATION_GROUP_ALERT_BEHAVIOR = 'groupAlertBehavior';
 const NOTIFICATION_ONLY_ALERT_ONCE = 'onlyAlertOnce';
 const NOTIFICATION_CHANNEL_KEY = 'channelKey';
@@ -175,6 +176,7 @@ const NOTIFICATION_ALLOW_WHILE_IDLE = 'allowWhileIdle';
 class Definitions {
   static Map<String, Object> initialValues = {
     NOTIFICATION_ID: 0,
+    NOTIFICATION_GROUP_SORT: GroupSort.Asc,
     NOTIFICATION_GROUP_ALERT_BEHAVIOR: GroupAlertBehavior.All,
     NOTIFICATION_IMPORTANCE: NotificationImportance.Default,
     NOTIFICATION_LAYOUT: NotificationLayout.Default,