+++ lib/src/enumerators/default_ringtone_type.dart
@@ -4,7 +4,10 @@
 /// [Notification]: Type that refers to sounds that are used for notifications (default).
 /// [Alarm]: Type that refers to sounds that are used for the alarm.
 enum DefaultRingtoneType {
+  /// Type that refers to sounds that are used for notifications (default).
   Notification,
+  /// Type that refers to sounds that are used for the phone ringer.
   Ringtone,
+  /// Type that refers to sounds that are used for the alarm.
   Alarm
 }